# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/albums/search": {
      "get": {
        "operationId": "AlbumCatalogController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumSearchPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
...
    "/v1/albums/new-releases": {
      "get": {
        "operationId": "AlbumCatalogController_newReleases",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "rock",
                "pop",
                "hip-hop",
                "r-b",
                "jazz",
                "electronic",
                "metal",
                "indie",
                "classical",
                "country",
                "reggae",
                "funk",
                "soul",
                "punk",
                "blues",
                "folk",
                "k-pop",
                "mpb",
                "samba",
                "sertanejo",
                "pagode"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewReleasesPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
...
    "/v1/albums/top-chart": {
      "get": {
        "operationId": "AlbumCatalogController_topChart",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopChartPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
...
    "/v1/albums/{spotifyId}": {
      "get": {
        "operationId": "AlbumCatalogController_getAlbum",
        "parameters": [
          {
            "name": "spotifyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
    "/v1/albums/{spotifyId}/cover": {
      "get": {
        "operationId": "AlbumCatalogController_getCover",
        "parameters": [
          {
            "name": "spotifyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bytes da capa (image/jpeg)."
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
    "/v1/albums/{spotifyId}/tracks/{trackId}/preview": {
      "get": {
        "operationId": "AlbumCatalogController_getTrackPreview",
        "parameters": [
          {
            "name": "spotifyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trackId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackPreviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
...
    "/v1/rankings/users/{userId}/albums/{albumId}": {
      "get": {
        "operationId": "RankingsController_getByUserAndAlbum",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "rankings"
        ]
      }
    },
...
    "/v1/rankings/{rankingId}/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "name": "rankingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
    "/v1/comments/{commentId}": {
      "patch": {
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentViewDto"
                }
              }
            }
          }
        },
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "comments"
        ]
      }
    }
  },
  "info": {
    "title": "Track by Track API",
    "description": "Contrato da API — fonte de tipos para o frontend (ver seção 2.2 do CLAUDE.md)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AlbumSearchPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "spotifyId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "artist": {
                  "type": "string"
                },
                "imageUrl": {
                  "type": "string"
                },
                "imageUrlSmall": {
                  "type": "string"
                },
                "releaseDate": {
                  "type": "string"
                }
              },
              "required": [
                "spotifyId",
                "name",
                "artist"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "NewReleasesPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "spotifyId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "artist": {
                  "type": "string"
                },
                "imageUrl": {
                  "type": "string"
                },
                "imageUrlSmall": {
                  "type": "string"
                },
                "releaseDate": {
                  "type": "string"
                },
                "genres": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "rock",
                      "pop",
                      "hip-hop",
                      "r-b",
                      "jazz",
                      "electronic",
                      "metal",
                      "indie",
                      "classical",
                      "country",
                      "reggae",
                      "funk",
                      "soul",
                      "punk",
                      "blues",
                      "folk",
                      "k-pop",
                      "mpb",
                      "samba",
                      "sertanejo",
                      "pagode"
                    ]
                  }
                }
              },
              "required": [
                "spotifyId",
                "name",
                "artist",
                "genres"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TopChartPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "spotifyId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "artist": {
                  "type": "string"
                },
                "imageUrl": {
                  "type": "string"
                },
                "releaseDate": {
                  "type": "string"
                },
                "genres": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "spotifyId",
                "name",
                "artist",
                "genres"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AlbumDetailDto": {
        "type": "object",
        "properties": {
          "spotifyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageUrlSmall": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string"
          },
          "tracks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "spotifyId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "durationMs": {
                  "type": "number"
                },
                "trackNumber": {
                  "type": "number"
                },
                "previewUrl": {
                  "type": "string"
                }
              },
              "required": [
                "spotifyId",
                "name",
                "durationMs",
                "trackNumber"
              ]
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "spotifyId",
          "name",
          "artist",
          "tracks"
        ]
      },
      "TrackPreviewDto": {
        "type": "object",
        "properties": {
          "previewUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "previewUrl"
        ]
      },
...
      "RankingViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "albumId": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "trackId": {
                  "type": "string"
                },
                "score": {
                  "type": "number"
                },
                "position": {
                  "type": "number"
                },
                "ignored": {
                  "type": "boolean"
                }
              },
              "required": [
                "trackId",
                "score",
                "position",
                "ignored"
              ]
            }
          },
          "averageScore": {
            "type": "number"
          },
          "progress": {
            "type": "object",
            "properties": {
              "rated": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "ignored": {
                "type": "number"
              },
              "percentage": {
                "type": "number"
              }
            },
            "required": [
              "rated",
              "total",
              "ignored",
              "percentage"
            ]
          },
          "review": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "favoriteTrackId": {
                "type": "string"
              },
              "worstTrackId": {
                "type": "string"
              }
            }
          },
          "isFirstCompletionBadgeActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "albumId",
          "entries",
          "averageScore",
          "progress",
          "review",
          "isFirstCompletionBadgeActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "text"
        ]
      },
      "CommentViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rankingId": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "authorDisplayName": {
            "type": "string"
          },
          "authorAvatarUrl": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "editedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "rankingId",
          "authorId",
          "authorDisplayName",
          "text",
          "createdAt",
          "editedAt"
        ]
      },
      "CommentsPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "rankingId": {
                  "type": "string"
                },
                "authorId": {
                  "type": "string"
                },
                "authorDisplayName": {
                  "type": "string"
                },
                "authorAvatarUrl": {
                  "type": "string"
                },
                "text": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                },
                "editedAt": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "rankingId",
                "authorId",
                "authorDisplayName",
                "text",
                "createdAt",
                "editedAt"
              ]
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "perPage": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "totalPages": {
                "type": "number"
              },
              "nextCursor": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "page",
              "perPage",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "text"
        ]
      }
    }
  }
}